Switch metrics up to be able to print concurrent request count via /healthz
In order to troubleshoot some observability issues we're seeing with the
current request metrics, this switches to using atomic.Int64 values to track
the current number of GET and POST requests in-flight. We can use these
values with an Int64ObservableUpDownCounter to send to our o11y platform,
but we can also print them out via the `/healthz` endpoint and use that for
troubleshooting whether the value we see in our dashboards is the same value
that the application is actually tracking.